Fix: motorITF would not set motor power if it was invalid
If motorITF was provided a power outside the range -127.0f -> 127.0f, the motor power would not be set. In a scenario where the motor power is invalid though, it is more likely one of the bounds were intended. E.x. a motor power of 128.0f is likely meant to be 127.0f. The change sets the motor power regardless of whether it is invalid or not.